+++ nkululeko/models/model_cnn.py
@@ -64,7 +64,7 @@ class CNNModel(Model):
             self.model.parameters(),
             self.util,
             default_lr=0.0001,
-            default_optimizer="adam",
+            default_optimizer="adamw",
         )
 
         # batch size