+++ nkululeko/models/model_mlp.py
@@ -72,7 +72,7 @@ class MLPModel(Model):
             self.model.parameters(),
             self.util,
             default_lr=0.0001,
-            default_optimizer="adam",
+            default_optimizer="adamw",
         )
 
         # batch size